+++ src/Mesh.jl
@@ -1,18 +1,24 @@
+using LinearAlgebra
+
 @enum BoundaryType NEUMANN=1 DIRICHLET=2 RADIATION=3 HEAT_FLUX=4
 
 mutable struct Boundary
     type::BoundaryType
     value
-    Boundary() = new(DIRICHLET, 0.0)
+    Boundary() = new(DIRICHLET, 1.0)
 end
 
 mutable struct Surface
-    thermal_conductivity
-    heat_capacity
-    density
-    volumetric_heat_source
+    properties::Dict{String, Number}
 
-    Surface() = new(1.0, 1.0, 1.0, 0.0)
+    function Surface()
+        properties::Dict{String, Number} = Dict()
+
+        properties["thermal_conductivity"] = 1.0
+        properties["volumetric_heat_source"] = 0.0
+
+        new(properties)
+    end
 end
 
 mutable struct Node
@@ -41,8 +47,40 @@ mutable struct Mesh
 
     boundary_names::Dict{String, Int}
     surface_names::Dict{String, Int}
+    boundary_ids::Dict{Int, Int}
+    surface_ids::Dict{Int, Int}
 
-    Mesh() = new([], [], [], [], Dict(), Dict())
+    Mesh() = new([], [], [], [], Dict(), Dict(), Dict(), Dict())
+end
+
+function set_boundary!(mesh, name, type, value)
+    bound_id = mesh.boundary_names[name]
+    boundary = mesh.boundaries[bound_id]
+    boundary.type = type
+    boundary.value = value
+end
+
+function set_surface_property!(mesh, name, type, value)
+    surf_id = mesh.surface_names[name]
+    surface = mesh.surfaces[surf_id]
+
+    if type in keys(surface.properties)
+        surface.properties[type] = value
+    else
+        throw(ErrorException("type [" * type * "] is not a surface property"))
+    end
+end
+
+function get_cell_area(mesh, cell_id)
+    cell = mesh.cells[cell_id]
+    point1 = mesh.nodes[cell.nodes[1]].position
+    point2 = mesh.nodes[cell.nodes[2]].position
+    point3 = mesh.nodes[cell.nodes[3]].position
+    dist1 = point2 - point1
+    dist2 = point3 - point1
+    cross_prod = cross(dist1, dist2)
+
+    return 0.5 * norm(cross_prod)
 end
 
 function read_Gmsh_file(file_name)
@@ -86,12 +124,14 @@ function setup_physical_properties!(mesh, gmsh_file)
 
             name_str = rstrip(lstrip(name[3], '"'), '"')
             mesh.boundary_names[name_str] = length(mesh.boundaries)
+            mesh.boundary_ids[parse(Int, name[2])] = length(mesh.boundaries)
         elseif length(name) == 3 && name[1] == "2"
             surface = Surface()
             push!(mesh.surfaces, surface)
 
             name_str = rstrip(lstrip(name[3], '"'), '"')
             mesh.surface_names[name_str] = length(mesh.surfaces)
+            mesh.surface_ids[parse(Int, name[2])] = length(mesh.surfaces)
         end
     end
 end
@@ -112,7 +152,8 @@ function setup_cells_and_nodes!(mesh, gmsh_file)
     for element in gmsh_file["Elements"]
         if length(element) > 1 && element[2] == "2"
             cell = Cell()
-            cell.surface_id = parse(Int, element[4])
+            surf_id = parse(Int, element[4])
+            cell.surface_id = mesh.surface_ids[surf_id]
             cell.nodes[1] = parse(Int, element[6])
             cell.nodes[2] = parse(Int, element[7])
             cell.nodes[3] = parse(Int, element[8])
@@ -128,11 +169,14 @@ function setup_cells_and_nodes!(mesh, gmsh_file)
                 items = findall(x -> x == parse(Int, element[6]) || x == parse(Int,element[7]), cell.nodes)
                 if length(items) == 2
                     if items[1] == 1 && items[2] == 2
-                        cell.boundaries[1] = parse(Int, element[4])
+                        id = parse(Int, element[4])
+                        cell.boundaries[1] = mesh.boundary_ids[id]
                     elseif items[1] == 2 && items[2] == 3
-                        cell.boundaries[2] = parse(Int, element[4])
+                        id = parse(Int, element[4])
+                        cell.boundaries[2] = mesh.boundary_ids[id]
                     elseif items[1] == 1 && items[2] == 3
-                        cell.boundaries[3] = parse(Int, element[4])
+                        id = parse(Int, element[4])
+                        cell.boundaries[3] = mesh.boundary_ids[id]
                     else
                         error("unacceptrable combination of cell indices")
                     end
@@ -153,45 +197,33 @@ function connect_mesh!(mesh)
     # sort adjacent cell
     for n=1:length(mesh.nodes)
         node = mesh.nodes[n]
-        if length(node.adjacent_cells) < 3
-            continue
-        end
-
-        c1 = 1
-        while true
-            if c1 >= (length(node.adjacent_cells) - 1)
+        for c=1:length(node.adjacent_cells)
+            cell = mesh.cells[node.adjacent_cells[c]]
+            pos = findall(x -> x == n, cell.nodes)[begin]
+            if sum(cell.boundaries) > -3 && cell.boundaries[pos] > -1
+                c_old = node.adjacent_cells[1]
+                node.adjacent_cells[1] = node.adjacent_cells[c]
+                node.adjacent_cells[c] = c_old
                 break
             end
+        end
 
-            id1 = node.adjacent_cells[c1]
+        for c_first=1:length(node.adjacent_cells)
+            id1 = node.adjacent_cells[c_first]
             pos_c1 = findall(x -> x == n, mesh.cells[id1].nodes)[begin]
             pos_c11 = pos_c1 > 1 ? pos_c1 - 1 : 3
             node_id1 = mesh.cells[id1].nodes[pos_c11]
-
-            for cc=(c1 + 1):length(node.adjacent_cells)
-                id2 = node.adjacent_cells[cc]
-                pos_c2 = findall(x -> x == n, mesh.cells[id2].nodes)[begin]
+            for c_second=(c_first+1):length(node.adjacent_cells)
+                id_second = node.adjacent_cells[c_second]
+                pos_c2 = findall(x -> x == n, mesh.cells[id_second].nodes)[begin]
                 pos_c22 =  pos_c2 < 3 ? pos_c2 + 1 : 1
-                node_id2 = mesh.cells[id2].nodes[pos_c22]
+                node_id2 = mesh.cells[id_second].nodes[pos_c22]
 
                 if node_id1 == node_id2
-                    c2_new = node.adjacent_cells[cc]
-                    node.adjacent_cells[cc] = node.adjacent_cells[c1 + 1]
-                    node.adjacent_cells[c1 + 1] = c2_new
-                    c1 += 1
+                    c2_new = node.adjacent_cells[c_second]
+                    node.adjacent_cells[c_second] = node.adjacent_cells[c_first + 1]
+                    node.adjacent_cells[c_first + 1] = c2_new
                     break
-                end
-
-                if cc == length(node.adjacent_cells)
-                    for q=1:length(node.adjacent_cells)
-                        if q!=c1 && sum(mesh.cells[node.adjacent_cells[q]].boundaries) > -3
-                            c_new = node.adjacent_cells[q]
-                            node.adjacent_cells[q] = node.adjacent_cells[1]
-                            node.adjacent_cells[1] = c_new
-                            c1 = 1
-                            break
-                        end
-                    end
                 end
             end
         end
@@ -209,6 +241,22 @@ function connect_mesh!(mesh)
             pos_p2 = pos > 1 ? pos - 1 : 3
             if !(cell.nodes[pos_p1] in node.adjacent_nodes) push!(node.adjacent_nodes, cell.nodes[pos_p1]) end
             if !(cell.nodes[pos_p2] in node.adjacent_nodes) push!(node.adjacent_nodes, cell.nodes[pos_p2]) end
+        end
+    end
+
+    # connect nodes to boundaries
+    for n=1:length(mesh.nodes)
+        node = mesh.nodes[n]
+        for c in node.adjacent_cells
+            cell = mesh.cells[c]
+            pos1 = findall(x -> x == n, cell.nodes)[1]
+            pos2 = pos1 == 1 ? 3 : pos1 - 1
+            if cell.boundaries[pos1] > -1
+                push!(node.boundaries, cell.boundaries[pos1])
+            end
+            if cell.boundaries[pos2] > -1
+                push!(node.boundaries, cell.boundaries[pos2])
+            end
         end
     end
 end