+++ src/Writer.jl
@@ -1,9 +1,10 @@
 using WriteVTK
 using HAMT
 
-function write_mesh(mesh)
+function write_mesh(file_name, mesh, solution)
     points::Matrix{Float64} = zeros(3, 3*length(mesh.cells))
     cells::Vector{MeshCell} = []
+    sol::Matrix{Float64} = zeros(1, 3*length(mesh.cells))
 
     for cell in mesh.cells
         push!(cells, MeshCell(VTKCellTypes.VTK_TRIANGLE, cell.nodes))
@@ -11,10 +12,11 @@ function write_mesh(mesh)
             points[1, node_id] = mesh.nodes[node_id].position[1]
             points[2, node_id] = mesh.nodes[node_id].position[2]
             points[3, node_id] = mesh.nodes[node_id].position[3]
+            sol[1, node_id] = solution[node_id]
         end
     end
 
-    vtk_grid("filename", points, cells) do vtk
-        # add datasets...
+    vtk_grid(file_name, points, cells) do vtk
+        vtk["temperature", VTKPointData()] = sol
     end
 end