+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-09-19
+
+- implemented solver module for 2D tringular mesh in cartesian coordinates
+- implemented Dirichlet boundary condition
+- implemented Neumann boundary condition
+
 ## [0.3.0] - 2024-09-16
 
 - implemented first vtk writing capabilities